Create 'remote' accessor
Now we're using Active Record Types to serialize to and deserialize from
the ID to the record. This should start to simplify things in a major
way.

The `remote_record` DSL method assumes you only have one remote
resource, linked at the `remote_resource_id` field on your model. The
aim is for it to be as easy as saying something like:

`attribute :github_user, RemoteRecord::GitHub::User::Type.new`.